requirements.txt
albumentations
aiohttp==3.11.16
backports.tarfile==1.2.0
bokeh==3.7.2
boto3==1.34.106
bottleneck==1.4.2
brotli==1.1.0
cartopy==0.24.1
cffi==1.17.1
conflator==0.1.7
crc32c==2.7.1
dask[array]==2024.5.2
distributed==2024.5.2
dotenv==0.9.9
exceptiongroup==1.2.2
flox==0.10.1
folium==0.19.5
geodatasets==2024.8.0
geopandas==1.0.1
gradio==5.23.3
h2==4.2.0
h5netcdf==1.6.1
importlib-metadata==6.8.0
ipykernel==6.29.5
isort==6.0.1
jaraco.collections==5.1.0
loguru==0.7.3
lxml==4.9.4
lz4==4.4.4
natsort==8.4.0
nbformat==5.10.4
nc-time-axis==1.4.1
netcdf4==1.7.2
numbagg==0.9.0
numpy==2.3.3
opt-einsum==3.4.0
plotly
pickleshare==0.7.5
pillow==11.3.0
pyarrow==19.0.1
pysocks==1.7.1
pystac-client==0.8.6
rasterio==1.3.10
seaborn==0.13.2
segmentation-models-pytorch
sentinelsat==1.2.1
sparse==0.16.0
tomli==2.0.1
torch==2.7.1
torchmetrics==1.7.4
torchvision==0.22.1
wandb==0.21.0
xarray==2024.05.0
zarr==2.18.4
zstandard==0.23.0